feat: getAnalysisBusiness

+++ src/main/java/com/kiosk/kioskback/dto/response/analysis/GetAnalysisSaleResponseDto.java
@@ -18,6 +18,6 @@ public class GetAnalysisSaleResponseDto {
     private int saleCount;
 
     @ApiModelProperty()
-    private double avgSaleAmount;
+    private int avgSaleAmount;
     
 }
